+++ ImpactfulSkills/ImpactfulSkills.cs
@@ -29,7 +29,7 @@ namespace ImpactfulSkills
     {
         public const string PluginGUID = "MidnightsFX.ImpactfulSkills";
         public const string PluginName = "ImpactfulSkills";
-        public const string PluginVersion = "0.8.0";
+        public const string PluginVersion = "0.8.1";
 
         public ValConfig cfg;
         // Use this class to add your own localization to the game